requirements.txt
langgraph>=0.1.0
langchain>=0.2.0
langchain-text-splitters>=0.2.0
langchain-groq>=0.1.0
langchain-community>=0.2.0
sentence-transformers>=3.0
chromadb>=0.5.0
requests>=2.31.0
beautifulsoup4>=4.12.0
duckduckgo-search>=6.0.0
python-dotenv>=1.0.0
rich>=13.0.0
